requirements.txt
pyqt6~=6.8.1
pyqt6-sip~=13.10.0
opencv-python~=4.11.0.86
mediapipe~=0.10.21
pillow~=11.1.0
jupyterlab-pygments~=0.3.0
pip~=25.0
attrs~=25.3.0
wheel~=0.45.1
protobuf~=4.25.6
tornado~=6.4.2
jinja2~=3.1.6